Remove common methods to mixin and fix gemfile issue.
